Refactored approach to fetching settings.
In many cases, the settings represent app-wide configurations for the AngularJS
app. It is convenient to make these available to the client side before
AngularJS bootstraps rather resolve them before each route/state where they are
needed.

The new approach allows "autodiscovery" of settings via the convention of
prefixing setting names with the shortname of the extension.