First pass at API permissions management.
This is a bit of a work in progress: several API calls remain unaddressed,
but we have a nice framework for adding new ones.

The nz.co.fuzion.relatedpermissions dependency was added because it gets
us some important stuff for free: namely, it allows an individual who has
a permissioned relationship over an organization to use the API to edit
said organization.