AngularJS page now determines the contact ID for the acting user and makes it available client-side.
The CiviCRM API Explorer pupports that passing "user_contact_id" as the value for parameters such
as api.Relationship.get's contact_id_a (in place of the usual integer) will autoselect the current
user. While this is true in many cases, user testing indicates that the same code behaves differently
depending on the permissions available to the acting user, so this technique can not be relied on.
Hence, this patch.